serving-endpoint/requirements.txt
databricks-sdk>=0.28.0
requests>=2.31.0

